+++ metadata.rb
@@ -4,12 +4,12 @@ maintainer_email 'me@ryangeyer.com'
 license          'All rights reserved'
 description      'Installs/Configures chef-server-blueprint'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.0.0'
+version          '1.0.3'
 
 depends "marker"
 depends "packagecloud"
 depends "chef-server", '= 5.0.1'
-depends "rsc_ros", "~> 0.3.0"
+depends "rsc_ros", "~> 0.7.0"
 
 # Support everything the chef-server cookbook supports
 %w{ ubuntu redhat centos fedora amazon scientific oracle }.each do |os|