+++ recipes/chef-ros-backup.rb
@@ -34,24 +34,33 @@ end
 # Is set as ENV['STORAGE_OPTIONS'] for ros_util.
 require 'json'
 
-backup_dir = '/var/backups/chef-backup'
-backup_src = File.join(backup_dir,'chef-backup.tar.bz2')
-backup_dst = File.join(backup_dir,node['chef-server-blueprint']['backup']['lineage']+ "-" + Time.now.strftime("%Y%m%d%H%M") + ".tar.bz2")
+backup_dir = '/var/opt/chef-backup'
+backup_src = File.join(backup_dir,'chef-backup.tgz')
+backup_file = node['chef-server-blueprint']['backup']['lineage']+ "-" + Time.now.strftime("%Y%m%d%H%M") + ".tgz"
+backup_dst = File.join(backup_dir,backup_file)
 
-bash "*** Uploading '#{backup_src}' to '#{cloud}' container '#{container}/chef-backups/#{backup_dst}'" do
+
+bash "Create backup file #{backup_src}'" do
   flags "-ex"
   user "root"
   code <<-EOH
     #{backup_script} --backup
-    mv #{backup_src} #{backup_dst}
   EOH
 end
 
-rsc_ros "uploading files" do
+ruby_block "Rename #{backup_src} to #{backup_dst}" do
+  block do
+    ::File.rename(backup_src,backup_dst)
+  end
+end
+
+log "Uploading '#{backup_src}' to '#{cloud}' container '#{container}/#{backup_file}"
+rsc_ros "Uploading backkup #{backup_dst}" do
   storage_provider  cloud
   access_key        node['chef-server-blueprint']['backup']['storage_account_id']
   secret_key        node['chef-server-blueprint']['backup']['storage_account_secret']
   bucket            container
   file              backup_dst
+  region            node['chef-server-blueprint']['backup']['region']
   action            :upload
 end